backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Backend for Nuntium EM"
authors = ["Raffay Rana <ranaraffay@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
gunicorn = "^23.0.0"
sqlalchemy = "^2.0.35"
httpx = "^0.27.2"
pycountry = "^24.6.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pymongo = "^4.9.1"
fastapi-users = {extras = ["beanie"], version = "13.0.0"}
pydantic-settings = "^2.5.2"
beanie = "^1.26.0"
redis = "^5.1.0"
fastapi-cache2 = "^0.2.2"
fastapi-mail = "^1.4.1"
slowapi = "^0.1.9"
requests = "^2.32.3"
yfinance = "^0.2.43"
mongomock-motor = "^0.0.34"
asgi-lifespan = "==2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["./app"]
asyncio_mode = "auto"
addopts = "--asyncio-mode=auto"